Move: move global search provider contract.
