--- packages/admin/src/GlobalSearch/Contracts/GlobalSearchProvider.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Filament\GlobalSearch\Contracts;
-
-use Filament\GlobalSearch\GlobalSearchResults;
-
-interface GlobalSearchProvider
-{
-    public function getResults(string $query): ?GlobalSearchResults;
-}
