+++ tests/src/Admin/GlobalSearch/GlobalSearchTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use Filament\Facades\Filament;
-use Filament\GlobalSearch\Contracts\GlobalSearchProvider;
+use Filament\Contracts\Search\GlobalSearchProvider;
 use Filament\GlobalSearch\GlobalSearchResult;
 use Filament\GlobalSearch\GlobalSearchResults;
 use Filament\Http\Livewire\GlobalSearch;
